feat: add POST /products endpoint

+++ src/main/resources/data.sql
@@ -2,5 +2,5 @@ CREATE TABLE product (
   id varchar(255) not null primary key,
   name varchar(50) not null,
   description varchar(400),
-  priceInCents numeric(18, 0)
+  priceInCents integer
 );
